+++ client/Packages/com.beamable.server/CHANGELOG.md
@@ -5,8 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+## [Unreleased]
+
  - Fix issue that was causing error while publishing microservice in apple silicon cpu architectures.
+
+## [1.19.9]
+
+### Fixed
+
+- Deserializing string field containing json in microservices compatibility improvements.
+
+## [1.19.8]
+
+no changes
 
 ## [1.19.7]